requirements.txt
# Core dependencies
pandas>=2.1.0
prophet>=1.1.0
numpy>=1.26.0
matplotlib>=3.8.0
db-dtypes

# Google Cloud dependencies
google-cloud-bigquery>=3.14.0
google-auth>=2.26.0
google-cloud-core>=2.4.0
pandas-gbq>=0.26.1

# Monarch Money API client
# Using community fork that supports new api.monarch.com endpoint
# (monarchmoney-enhanced doesn't support the January 2026 domain change)
monarchmoneycommunity>=1.1.0
gql>=3.4.0

# Async dependencies
asyncio>=3.4.3
aiohttp>=3.11.0

# Utilities
python-dateutil>=2.8.0
python-dotenv>=1.0.0
